Change int64 to int for count-related fields and methods
This commit updates various fields, methods, and test cases to use `int` instead of `int64` for count-related variables. This simplifies type handling and aligns with standard conventions where 64-bit precision is unnecessary. All affected components have been updated consistently to reflect this change.

+++ internal/backends/docker/function.go
@@ -10,8 +10,8 @@ type Function struct {
 	Name_    string            `json:"name"`
 	Image_   string            `json:"image"`
 	Timeout_ time.Duration     `json:"timeout"`
-	MinScale int64             `json:"minScale"`
-	MaxScale int64             `json:"maxScale"`
+	MinScale int               `json:"minScale"`
+	MaxScale int               `json:"maxScale"`
 	Env_     map[string]string `json:"env"`
 }
